re-arrange files to split loading
researcher/participant shouldn't need to load each others views
