+++ frontend/src/models/study.ts
@@ -67,6 +67,7 @@ export const StudyValidationSchema = Yup.object().shape({
 
 export const LaunchStudy = async (api: StudiesApi, study: {id: number}, options: { preview?: boolean } = {}) => {
     const launch = await api.launchStudy({ id: study.id, preview: options.preview || false })
+    debugger
     window.location.assign(launch.url!)
     return launch
 }
